third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc events with invalidations.

PASS
PASS
PASS


Running: testLocalFrame
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:9}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:10}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: multipleStyleRecalcs
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:21}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:23}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:24}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:26}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:27}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:28}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementThree'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:35}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../devtools/tracing/timeline-style/timeline-style-recalc-with-invalidations.html:36}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
